backend/requirements.txt
fastapi
pydantic
uvicorn
asyncio
asyncpg
httpx
datetime
typing
pydantic[email]
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib
bcrypt
python-multipart
alembic
pydantic-settings
sqlalchemy[asyncio]
bcrypt
